# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Application API",
        "description": "API Documentation for Online Application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/placeOrder": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "تأكيد الطلب",
                "operationId": "21dbb34358712a9e97ec41b1f266b281",
                "responses": {
                    "201": {
                        "description": "تم طلب الأوردر بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "تم طلب الأوردر بنجاح"
                                        },
                                        "order_id": {
                                            "type": "integer",
                                            "example": 55
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "الكارت فاضي أو منتج out of stock",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "الكارت فاضي"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/getOrderDetails/{order_id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "تفاصيل أوردر معين",
                "operationId": "f4c57c2cd7b5ca28cd0afab27f08f86c",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 55
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تفاصيل الأوردر",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "order_id": {
                                                    "type": "integer",
                                                    "example": 55
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "placed"
                                                },
                                                "final_price": {
                                                    "type": "number",
                                                    "example": 345
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-01-01 10:00:00"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "image": {
                                                                "type": "string",
                                                                "example": "http://example.com/storage/products/img.jpg",
                                                                "nullable": true
                                                            },
                                                            "product_id": {
                                                                "type": "integer",
                                                                "example": 101
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "منتج A"
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "unit_price": {
                                                                "type": "number",
                                                                "example": 50
                                                            },
                                                            "unit_tax": {
                                                                "type": "number",
                                                                "example": 7.5
                                                            },
                                                            "unit_price_after_tax": {
                                                                "type": "number",
                                                                "example": 57.5
                                                            },
                                                            "total_price": {
                                                                "type": "number",
                                                                "example": 115
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "الأوردر مش موجود",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "الأوردر مش موجود"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/getUserOrdersHistory": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "سجل الطلبات بتاع اليوزر",
                "operationId": "24e75b4ad611c4254e9ecc099917cdbf",
                "responses": {
                    "200": {
                        "description": "قائمة الطلبات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "order_id": {
                                                        "type": "integer",
                                                        "example": 55
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "placed"
                                                    },
                                                    "total_price": {
                                                        "type": "number",
                                                        "example": 345
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-01-01 10:00:00"
                                                    },
                                                    "items": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "image": {
                                                                    "type": "string",
                                                                    "nullable": true
                                                                },
                                                                "product_id": {
                                                                    "type": "integer",
                                                                    "example": 101
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "منتج A"
                                                                },
                                                                "quantity": {
                                                                    "type": "integer",
                                                                    "example": 2
                                                                },
                                                                "unit_price": {
                                                                    "type": "number",
                                                                    "example": 50
                                                                },
                                                                "unit_tax": {
                                                                    "type": "number",
                                                                    "example": 7.5
                                                                },
                                                                "unit_price_after_tax": {
                                                                    "type": "number",
                                                                    "example": 57.5
                                                                },
                                                                "total_price": {
                                                                    "type": "number",
                                                                    "example": 115
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cancelOrder/{order_id}": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "إلغاء أوردر",
                "operationId": "6c8becb8e6925f10cb933dfe37e98ea0",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 55
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم إلغاء الأوردر بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "تم إلغاء الأوردر بنجاح"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "مش ممكن تلغي الأوردر ده",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "مش ممكن تلغي الأوردر ده"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "الأوردر مش موجود",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "الأوردر مش موجود"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}